move discrete_choice fixtures into their own file
